==> 9. Машинное обучение в бизнесе/9. Интеграция. Итоговый проект/step_2.ipynb <==
### Проверка работоспособности и качества пайплайна

Здесь мы еще не запускаем никакое API, а загружаем модель (pipeline) напрямую и проверяем на отложенной (тестовой) выборке

import pandas as pd
from sklearn.metrics import roc_auc_score, roc_curve
import import_ipynb
from step_1 import data_preprocessing, target_preprocessing#, stations
import dill
dill._dill._reverse_typemap['ClassType'] = type

X_test = pd.read_csv("X_test.csv")
y_test = pd.read_csv("y_test.csv")

X_test = data_preprocessing(X_test)
y_test = target_preprocessing(y_test)

#save test
X_test.to_csv("X_test.csv", index=None)
y_test.to_csv("y_test.csv", index=None)

with open('gbc_pipeline.dill', 'rb') as in_strm:
    pipeline = dill.load(in_strm)

predictions = pipeline.predict_proba(X_test)
pd.DataFrame({'preds': predictions[:, 1]}).to_csv("test_predictions.csv", index=None)

roc_auc_score(y_score=predictions[:, 1][:], y_true=y_test.iloc[:])
