==> 9. Машинное обучение в бизнесе/9. Интеграция. Итоговый проект/Step_3.ipynb <==
Строим запрос к нашему сервису

import pandas as pd
import requests
import json
from urllib import request, parse
from sklearn.metrics import roc_auc_score, roc_curve

# Пример данных
data = (
    '2009-03-20', #Date
    'Darwin', #Location
    24.6, #MinTemp
    34.7, #MaxTemp
    1e-8, #Rainfall
    8.4, #Evaporation
    11.1, #Sunshine
    'W', #WindGustDir
    37.0, #WindGustSpeed
    'SSW', ##WindDir9am
    'WSW', #WindDir3pm
    17.0,#WindSpeed9am
    24.0, #WindSpeed3pm
    63.0, #Humidity9am
    29.0,#Humidity3pm
    1009.6, #Pressure9am
    1006.3, #Pressure3pm
    6.0, #Cloud9am
    3.0, #Cloud3pm
    29.7, #Temp9am
    34.6, #Temp3pm
    'No' #RainToday
)

# формируем запрос
def send_json(x):

    Date, Location, MinTemp, MaxTemp, Rainfall, Evaporation, Sunshine, \
    WindGustDir, WindGustSpeed, WindDir9am, WindDir3pm, WindSpeed9am, \
    WindSpeed3pm, Humidity9am, Humidity3pm, Pressure9am, Pressure3pm, \
    Cloud9am, Cloud3pm, Temp9am, Temp3pm, RainToday = x
    
    body = {
        "Date": Date,
        "Location": Location,
        "MinTemp": MinTemp,
        "MaxTemp": MaxTemp,
        "Rainfall": Rainfall,
        "Evaporation": Evaporation,
        "Sunshine": Sunshine,
        "WindGustDir": WindGustDir,
        "WindGustSpeed": WindGustSpeed,
        "WindDir9am": WindDir9am,
        "WindDir3pm": WindDir3pm,
        "WindSpeed9am": WindSpeed9am,
        "WindSpeed3pm": WindSpeed3pm,
        "Humidity9am": Humidity9am,
        "Humidity3pm": Humidity3pm,
        "Pressure9am": Pressure9am,
        "Pressure3pm": Pressure3pm,
        "Cloud9am": Cloud9am,
        "Cloud3pm": Cloud3pm,
        "Temp9am": Temp9am,
        "Temp3pm": Temp3pm,
        "RainToday": RainToday
    }
    myurl = 'http://127.0.0.1:5000/'+'predict'
    headers = {'content-type': 'application/json; charset=utf-8'}
    response = requests.post(myurl, json=body, headers=headers)
    return response.json()['predictions']

# обращение к серверу с запросом из одного набора (его построили руками выше - data)
if __name__ == '__main__':
    response = send_json(data)
    print('предсказание',response)

Сделаем обработку массы запросов:

1. загрузим данные X_test

X_test = pd.read_csv("X_test.csv")
y_test = pd.read_csv("y_test.csv")
X_test.head()

2. Заменим нули на очень малые значения

X_test = X_test.replace({0.0: 1e-8})

2. Передаем запрос одиночный из наших данных и посмотрим на ответ

X_test.iloc[8,:]

response = send_json(X_test.iloc[8,:])
response

response

X_test.head()

3. Сделаем N запросов и оценим время

N = 50

%%time
predictions = X_test.iloc[:N,:].apply(lambda x: send_json(x),1)

predictions.values[:5]

4. Посчитаем метрику roc-auc

y_test.head()

roc_auc_score(y_score=predictions.values, y_true=y_test.values[:N])

5. Анализ : в целом похоже
